+++ tests/Integration/DSC_SqlDatabase.Integration.Tests.ps1
@@ -287,6 +287,56 @@ try
             }
         }
 
+        $configurationName = "$($script:dscResourceName)_AddDatabase6_Config"
+
+        Context ('When using configuration {0}' -f $configurationName) {
+            It 'Should compile and apply the MOF without throwing' {
+                {
+                    $configurationParameters = @{
+                        OutputPath           = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData    = $ConfigurationData
+                    }
+
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path         = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait         = $true
+                        Verbose      = $true
+                        Force        = $true
+                        ErrorAction  = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
+                } | Should -Not -Throw
+            }
+
+            It 'Should be able to call Get-DscConfiguration without throwing' {
+                {
+                    $script:currentConfiguration = Get-DscConfiguration -Verbose -ErrorAction Stop
+                } | Should -Not -Throw
+            }
+
+            It 'Should have set the resource and all the parameters should match' {
+                $resourceCurrentState = $script:currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName `
+                    -and $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.Ensure | Should -Be 'Present'
+                $resourceCurrentState.Name | Should -Be $ConfigurationData.AllNodes.DatabaseName6
+                $resourceCurrentState.ServerName | Should -Be $ConfigurationData.AllNodes.ServerName
+                $resourceCurrentState.InstanceName  | Should -Be $ConfigurationData.AllNodes.InstanceName
+                $resourceCurrentState.SnapshotIsolation | Should -Be $ConfigurationData.AllNodes.SnapshotIsolation
+            }
+
+            It 'Should return $true when Test-DscConfiguration is run' {
+                Test-DscConfiguration -Verbose | Should -Be 'True'
+            }
+        }
+
         $configurationName = "$($script:dscResourceName)_RemoveDatabase2_Config"
 
         Context ('When using configuration {0}' -f $configurationName) {
@@ -333,6 +383,7 @@ try
                 $resourceCurrentState.OwnerName | Should -BeNullOrEmpty
                 $resourceCurrentState.RecoveryModel | Should -BeNullOrEmpty
                 $resourceCurrentState.CompatibilityLevel | Should -BeNullOrEmpty
+                $resourceCurrentState.SnapshotIsolation | Should -BeNullOrEmpty
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {